# src/openapi.yaml
openapi: 3.0.0
info:
  title: Verificable Credential Service Documentation
  description: Verificable Credential Service Api Documentation
  version: "0.1.0"
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api
    description: Default Server
  - url: "{protocol}://{environment}:{port}/api"
    variables:
      protocol:
        enum:
          - http
          - https
        default: https
      port:
        default: "443"
      environment:
        default: localhost
    description: Custom Server
...
paths:
...
  /auth/direct_post:
    post:
      summary: Direct Post Endpoint
      description: Endpoint for receiving the ID token in an x-www-form-urlencoded payload and returning a redirect with the code and state parameters.
      operationId: directPost
      tags:
        - Authorization
      requestBody:
        description: ID token payload
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                issuerUri:
                  description: The issuer URI
                  type: string
                  example: "https://my.example.issuer/issuer"
                id_token:
                  type: string
                  description: ID Token signed by the DID document's authentication key. This will be used to prove the control of the DID.
                vp_token:
                  type: string
                  description: VP Token signed by the DID document's authentication key.
                presentation_submission:
                  type: string
              required:
                - issuerUri
      responses:
        200:
          description: Successfully processed the token with no redirection.
          content: # Revisar. El 200 creo no debe devolver nada, pero si no lo declaro y no envío algo falla el backend al procesarlo
            application/json:
              schema:
                type: object
        302:
          description: Successfully processed the token and redirecting.
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    type: string
                    example: "openid://?code=SplxlOBeZQQYbYS6WxSbIA&state=client-state"
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /credentials/status:
    post:
      summary: Credentials Status
      description: Get the credential status
      operationId: statusCredentialRequest
      tags:
        - Credentials
      requestBody:
        description: Credential request parameters
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                issuerDid:
                  description: The issuer DID
                  type: string
                  example: "did:ebsi:zvHWX359A3CvfJnCYaAiAde"
                issuerUri:
                  description: The issuer URI
                  type: string
                  example: "https://my.example.issuer/issuer"
                listId:
                  description: The ID of the list
                  type: string
                  example: "/credentials/status/list/1"
                statusList:
                  description: Status List
                  type: string
                statisPurpose:
                  description: Status Purpose
                  type: string
                  enum:
                    - revocation
                    - suspension
                revocationType:
                  description: Revocation Type
                  type: string
                  enum:
                    - StatusList2021
      responses:
        200:
          description: Successfully issued credential
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/CredentialResponse"
                  - $ref: "#/components/schemas/DeferredCredentialResponse"
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  # EBSI Accreditation
  /ebsi/accreditation/issuance:
    get:
      summary: EBSI Accreditations direct issuance
      description: Allows the issuance of an EBSI Accreditation
      operationId: ebsiDirectAccreditationIssuance
      tags:
        - Credentials
      parameters:
        - name: holderDid
          in: query
          required: true
          description: The DID of the VC to issue
          schema:
            type: string
        - name: issuerDid
          in: query
          required: true
          description: The DID of the VC to issue
          schema:
            type: string
        - name: issuerUri
          in: query
          required: true
          description: The DID of the VC to issue
          schema:
            type: string
        - name: accreditationType
          in: query
          required: true
          description: The specific type of the accreditation
          schema:
            type: string
            enum:
              - VerifiableAccreditationToAccredit
              - VerifiableAccreditationToAttest
              - VerifiableAuthorisationToOnboard
      responses:
        # Not Standard
        200:
          description: VC Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  credential:
                    type: string
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  securitySchemes:
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CredentialResponse:
      type: object
      additionalProperties: false
      properties:
        format:
          type: string
          example: jwt_vc
        credential:
          type: string
          example: "LUpixVCWJk0eOt4CXQe1NXK....WZwmhmn9OQp6YxX0a2L"
        c_nonce:
          type: string
          example: "fGFF7UkhLa"
        c_nonce_expires_in:
          type: integer
          example: 86400
      required:
        - format
        - credential
    DeferredCredentialResponse:
      type: object
      additionalProperties: false
      properties:
        acceptance_token:
          type: string
          example: eyJ0eXAiOiJKV1QiLCJhbGci..zaEhOOXcifQ
        c_nonce:
          type: string
          example: "fGFF7UkhLa"
        c_nonce_expires_in:
          type: integer
          example: 86400
      required:
        - acceptance_token
...
    Error:
      type: object
      properties:
        error:
          type: string
        error_description:
          type: string